+++ go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/flytam/filenamify v1.2.0
 	github.com/spf13/cobra v1.10.2
 	gopkg.in/yaml.v3 v3.0.1
+	github.com/go-chi/chi/v5 v5.2.3
 )
 
 require (
