+++ server/router.go
@@ -1,0 +1,26 @@
+package server
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/go-chi/chi/v5"
+	"github.com/go-chi/chi/v5/middleware"
+)
+
+func NewRouter() http.Handler {
+	r := chi.NewRouter()
+	r.Use(middleware.Logger, middleware.Recoverer)
+
+	r.Route("/api", func(r chi.Router) {
+		r.Post("/tasks", createTask)
+		r.Get("/tasks", listTasks)
+		r.Get("/tasks/{vid}", getTask)
+		r.Delete("/tasks/{vid}", deleteTask)
+	})
+	return r
+}
+
+func RunServer(port int) error {
+	return http.ListenAndServe(fmt.Sprintf(":%d", port), NewRouter())
+}
